requirements.txt
# API & runtime
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
requests==2.32.3

# Core RAG stack (CPU-only, no Torch)
numpy==1.26.4          # py3.9-compatible
scikit-learn==1.4.2
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2

# Ingestion & parsing
pypdf==4.3.1
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0

# UI & reporting
streamlit==1.37.0
reportlab==4.2.0
markdown-it-py==3.0.0
tqdm==4.66.4

# Optional/utility
pandas==2.2.2
